changed all monetary values in every class from double to BigDecimal. Modified calcMoneyInHold method for to work with BigDecimal values. changed relevant tests to work with BigDecimal values. made makeChange method return the difference between the price of the selectedItem and the total of calcMoneyInHold.

+++ src/main/java/com/jchaffin/vendendingmachine/Item.java
@@ -1,10 +1,12 @@
 package com.jchaffin.vendendingmachine;
 
+import java.math.BigDecimal;
+
 public interface Item {
 
-	public void setPrice(double price);
+	public void setPrice(BigDecimal price);
 
-	public double getPrice();
+	public BigDecimal getPrice();
 
 	public void setType(String type);
